Merge resourceable into linked_payload
Also, don't check customize the `new` method to fail on invalid
initializations. It's possible to create an object that passes on
initialization but later fails (eg. by mutation, or by using a `field`
block that can change). For example, this caused problems:

    class R
      include LinkedPayload
      field :one, 1, -> { rand 2 }
    end

    R.new.fields

`R.new` would fail initialization half the time. And when it worked, the
`field` isn't cached, so `fields` would fail half the time.