Change `Decoder`'s verb to `test`
Seems more appropriate than `check`. Also, it's shorter.

I also considered using the verb `decode`, but that seems more
appropriate for decoding `JSON`.

I checked Elm's decoder object to see if they set a precedent I could
follow, but the decoders there are higher-order functions, which aren't
so common in Ruby.

I considered using the method `#[]`, so usage would be:

    is_string = Decoder.new { |unknown| unknown.is_a(String)? || "Err" }
    is_string["This is a string"] #=> ok("This is a string")
    is_string[:this_is_a_symbol]  #=> err("Err")

Which kind of looks nice, but it's very unusual, and I don't want to do
something so unexpected.